Add universal cache-handler and json-replacer-reviver packages, refactor server and tests
This commit adds the `@neshca/cache-handler` and `@neshca/json-replacer-reviver` packages to the project. These packages provide a universal cache handler and functions for working with `Buffer` objects and JSON. The server code has been refactored to use the new cache handler, and the `handler` and `handler-redis` packages have been removed.

In addition, the tests have been stabilized to ensure that they pass consistently.