Add ability to disable Redis operations timeout in the `redis-string` and `redis-stack` Handlers  (#639)
There is a known bug in the Redis package, that will be fixed in version 5, which is in the alpha stage. (https://github.com/redis/node-redis/issues/2498)

The issue is about the usage of AbortSignal. When it is used, if the Redis connection encounters an error, like the Redis server gets offline, an unhandled error will be thrown. This error will cause an issue with the reconnection of the client, and it'll put the Redis client into a limbo stage that will not recover unless the Next.js app restarts.

This PR will address this issue by giving the ability to disable the Redis operations timeout totally, by setting the value of `timeoutMs: 0` for the `redis-string` and `redis-stack` Handlers.