Implement pre-populating the cache with the initial data (#743)
This PR delivers the highly anticipated feature of pre-populating the cache with the initial data at application startup, resulting in improved performance. It introduces a new `registerInitialCache` function that takes a `CacheHandler` class (not an instance) and uses it to populate all cache stores.

We must do this process after the build to guarantee that the initial cache is generated and the `prerender-manifest.json` is created. The instrumentation hook allows us to run a function just before the server starts, providing the perfect moment to populate the cache. By using this approach, we eliminate the necessity for any additional JS or shell scripts.

_Note: This is an experimental feature and might change in future releases. To use it, you'll need to enable the `experimental.instrumentationHook` option in your `next.config.js`._

Fix #514